Code Cleanup
- Guaranteed the correct Config::get call via cast.  There is some
  doubt as to the correct call being made otherwise.  This is not a
  very likely fix, but does eliminate one possibility.

+++ src/command.cpp
@@ -283,7 +283,7 @@ std::string handleVersion (Config& conf)
   std::stringstream out;
 
   // Determine window size, and set table accordingly.
-  int width = conf.get ("defaultwidth", 80);
+  int width = conf.get ("defaultwidth", (int) 80);
 #ifdef HAVE_LIBNCURSES
   if (conf.get ("curses", true))
   {
